Try to make search best move modules more independent
